+++ src/Resources/Databases.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace R64\PhpNotion\Resources;
 
-use R64\PhpNotion\Integrations\NotionClient;
+use R64\PhpNotion\Client\NotionClient;
 
 class Databases
 {
